Fix forceRoundTime with minTime glitch
If forceRoundTime is true, and a step is given, and minTime is set, an inconsistency can occur when minTime is not a multiple of step from midnight.  For example, consider the case of a minTime of 6:45am and a step of 6. The documentation says "Setting forceRoundTime to true will round the entered time to the nearest option on the dropdown list."  However, in the case given, the dropdown will have 6:45am, 6:51am, 6:57am, ... but the initial value will set to 6:48am and if I type in, say, 6:51am it will jump to 6:54am which isn't even one of the dropdown options.